Improved helpers
